Struct radix_heap::RadixHeapMap

pub struct RadixHeapMap<K, V> {
    // some fields omitted
}

A montone priority queue implemented with a radix heap.

This will be a max-heap.

It is a logic error for a key to be modified in such a way that the item's ordering relative to any other item, as determined by the Ord trait, changes while it is in the heap. This is normally only possible through Cell, RefCell, global state, I/O, or unsafe code.

Methods

impl<K: Radix + Ord + Copy + Bounded, V> RadixHeapMap<K, V>

fn new() -> RadixHeapMap<K, V>

Creates a new RadixHeapMap with the highest possible top value such that all keys are allowed.

fn clear(&mut self)

Clears a new RadixHeapMap and sets te top value as high as possible value, so all keys can be inserted into the heap.

impl<K: Radix + Ord + Copy, V> RadixHeapMap<K, V>

fn new_at(top: K) -> RadixHeapMap<K, V>

Creates a new RadixHeapMap with a specific top value

fn clear_to(&mut self, top: K)

Clears the elements in the heap and sets the top value to a new value

fn push(&mut self, key: K, value: V)

Pushes a new key value pair onto the heap.

Panics

Panics if the key is more than the current top value.

fn constrain(&mut self)

Sets the top value to the current maximum key value in the heap

fn pop(&mut self) -> Option<(K, V)>

Pops the largest element of the heap. This might decrease the top value.

fn len(&self) -> usize

Returns the number of elements in the heap

fn is_empty(&self) -> bool

Returns true if there is no elements in the heap

fn top(&self) -> K

The current top value. All keys pushed onto the heap must be smaller than this value.

fn shrink_to_fit(&mut self)

Discards as much additional capacity as possible.