Struct radix_heap::RadixHeapMap

pub struct RadixHeapMap<K, V> { /* private fields */ }

A montone priority queue implemented using a radix heap.

This will be a max-heap.

See the module documentation for more information.

It is a logic error for a key to be modified in such a way that the item’s ordering relative to any other item, as determined by the Ord trait, changes while it is in the heap. This is normally only possible through Cell, RefCell, global state, I/O, or unsafe code.

Implementations

impl<K: Radix + Ord + Copy, V> RadixHeapMap<K, V>

pub fn new() -> RadixHeapMap<K, V>

Create an empty RadixHeapMap

pub fn new_at(top: K) -> RadixHeapMap<K, V>

Create an empty RadixHeapMap with the top key set to a specific value.

This can be more efficient if you have a known minimum bound of the items being pushed to the heap.

pub fn clear(&mut self)

Drops all items form the RadixHeapMap and sets the top key to None.

pub fn clear_to(&mut self, top: K)

Drop all items from the RadixHeapMap and sets the top key to a specific value.

This can be more efficient if you have a known maximum bound of the items being pushed to the heap.

pub fn constrain(&mut self)

Sets the top value to the current maximum key value in the heap

pub fn push(&mut self, key: K, value: V)

Pushes a new key value pair onto the heap.

Panics

Panics if the key is larger than the current top key.

pub fn pop(&mut self) -> Option<(K, V)>

Remove the greatest element from the heap and returns it, or None if empty.

If there is a tie between multiple elements, the last inserted element will be popped first.

This will set the top key to the extracted key.

pub fn len(&self) -> usize

Returns the number of elements in the heap

pub fn is_empty(&self) -> bool

Returns true if there is no elements in the heap

pub fn top(&self) -> Option<K>

The current top value. All keys pushed onto the heap must be smaller than this value.

pub fn shrink_to_fit(&mut self)

Discards as much additional capacity as possible.

pub fn iter(&self) -> Iter<'_, K, V>

Returns an iterator of all key-value pairs in the RadixHeapMap in arbitrary order

pub fn keys(&self) -> Keys<'_, K, V>

Returns an iterator of all keys in the RadixHeapMap in arbitrary order

pub fn values(&self) -> Values<'_, K, V>

Returns an iterator of all values in the RadixHeapMap in arbitrary order